feat(lesson5): use `range()` up to *length* var
Write a for loop with i as the loop variable. Use the `range()`
function to iterate up to the value of the length.
Inside the loop, use the addition assignment operator to add a random
character from all_characters to the current value of password. Use the
choice() function from the secrets module for that.