refactor(lesson7): replace `print()` with `if`
The allowed disk movements between the rods exhibit a repetitive pattern occurring every three moves. For example, movements between rod A and rod C are allowed on the first, the fourth and the seventh move, where the remainder of the division between the move number and 3 is equal to 1.

- Replace existing `print()` call with an `if` statement that is
  triggered when `(i + 1) % 3 == 1`
- Within `if` statement print `f'Move {i + 1} allowed between {source}
  and {target}'`

Note: `i + 1` is the move number since `i` is zero during the first
iteration